backend/requirements.txt
# Backend Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.6.0

# Neo4j
neo4j==5.15.0

# Google Gemini
google-generativeai==0.3.2

# Document processing (for scripts)
docling==1.9.0
pdf2image==1.17.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7
sentence-transformers>=2.2.0
